Refactor : Registration Screen from XML to Compose
